pyproject.toml
[tool.poetry]
name = "llmovies"
version = "0.1.0"
description = ""
authors = ["jlondonobo <j.londonobo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.9.1"
ruff = "^0.0.289"
pip = "^23.2.1"
install = "^1.3.5"
weaviate-client = "^3.24.1"
python-dotenv = "^1.0.0"
langchain = "^0.0.287"
openai = "^0.28.0"
networkx = "^3.1"
pandas = "^2.1.0"
sentence-transformers = "^2.2.2"
datasets = "^2.14.5"
streamlit = "^1.26.0"
cinemagoer = "^2023.5.1"
tmdbv3api = "^1.9.0"
asyncio = "^3.4.3"
aiohttp = "^3.8.5"
pydantic = "^2.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
